cleaned up maintsx file and added seperate authenticator component file
